+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add PushHistoriesService exposing api.touch.pushes.histories for touch/push-histories resource. @vhidvz
+
 ## [1.3.12] - 2026-05-02
 
 ## [1.3.11] - 2026-05-02